[PARAM, IO] OPXL: changed order and argument descriptions of output files
The order and descriptions are more in line with XFDR now
and the new order reflects the likely usage of the output formats much better
with idXML at the top. This will also make it more obvious that idXML is the best format to use in KNIME workflows, since it will be the top output of the node.